+++ src/components/MdToJson/MdToJson.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+export default function MdToJson() {
+  return <></>;
+}
+
+MdToJson.propTypes = {
+  /** string in Markdown format ('md' ) */
+  markdown: PropTypes.string,
+};