+++ src/components/index.js
@@ -1,1 +1,5 @@
-export { default as Button } from './Button';
+export { default as MdToJson } from './MdToJson';
+export { default as JsonToMd } from './JsonToMd';
+export { default as JsonToHtml } from './JsonToHtml';
+export { default as JsonToPdf } from './JsonToPdf';
+export { default as MdToZip } from './MdToZip';