+++ styleguide.config.js
@@ -8,8 +8,14 @@ let sections = [
     content: 'README.md',
   },
   {
-    name: 'Button Block',
-    components: ['src/components/Button/Button.js'],
+    name: 'Conversion block',
+    components: [
+      'src/components/MdToJson/MdToJson.jsx',
+      'src/components/JsonToMd/JsonToMd.jsx',
+      'src/components/JsonToHtml/JsonToHtml.jsx',
+      'src/components/JsonToPdf/JsonToPdf.jsx',
+      'src/components/MdToZip/MdToZip.jsx',
+    ],
   },
 ];
 
@@ -25,7 +31,7 @@ module.exports = {
       meta: [
         {
           name: 'description',
-          content: 'React component library template',
+          content: 'React component library for formatting and converting OBS data',
         },
       ],
     },
@@ -68,7 +74,7 @@ module.exports = {
   usageMode: 'expand',
   pagePerSection: true,
   getComponentPathLine(componentPath) {
-    const componentName = path.basename(componentPath, '.js');
+    const componentName = path.basename(componentPath, '.jsx');
     return `import { ${componentName} } from '${name}';`;
   },
   updateExample(props, exampleFilePath) {
@@ -89,7 +95,7 @@ module.exports = {
     module: {
       rules: [
         {
-          test: /\.js$/,
+          test: /\.jsx?$/,
           exclude: /node_modules/,
           loader: 'babel-loader',
         },