+++ CHANGELOG.md
@@ -2,4 +2,41 @@
 
 This repository is part of the open components ecosystem.
 
-### 0.1.1 (2023-01-24)
+## [0.3.0](https://github.com/texttree/obs-format-convert-rcl/compare/v0.1.2...v0.3.0) (2023-07-05)
+
+### Features
+
+- :sparkles: show/hide TitlePage ([4329c89](https://github.com/texttree/obs-format-convert-rcl/commit/4329c892f8c9009e6dce105087ffe08c66240e02))
+
+### Bug Fixes
+
+- :bug: if we don't pass styles to jsontopdf, the component broke, corrected ([3015b40](https://github.com/texttree/obs-format-convert-rcl/commit/3015b404e91bf7e70a6abe93954e9160fbf001a9))
+- :bug: line breaks fixed ([9b509d2](https://github.com/texttree/obs-format-convert-rcl/commit/9b509d20651c1cdf1408442bf01ee0c3678cfee7))
+
+### 0.1.1 (2023-06-22)
+
+### Features
+
+- :sparkles: added a parameter that disables the display of images in PDF ([3309925](https://github.com/texttree/obs-format-convert-rcl/commit/3309925a6e4d66d22750d9ec5abba180f15070ca))
+- :sparkles: added a parameter that says that the process of rendering pdf is in progress ([f18c09c](https://github.com/texttree/obs-format-convert-rcl/commit/f18c09c36aed5f54a5242fc697fa22eda586ae12))
+- :sparkles: added combineVerses ([82098f6](https://github.com/texttree/obs-format-convert-rcl/commit/82098f6160227558420e1545c4bc03b9af98ab0c))
+- :sparkles: added json-to-html ([437ad7c](https://github.com/texttree/obs-format-convert-rcl/commit/437ad7c069eb86273165bb912324abc0e0a159a0))
+- :sparkles: added json-to-md ([b8c0a05](https://github.com/texttree/obs-format-convert-rcl/commit/b8c0a05b01b752c1551836e5088fa447ae926f86))
+- :sparkles: added json-to-pdf ([0253f73](https://github.com/texttree/obs-format-convert-rcl/commit/0253f736f1893405284caf2161a497128b2938f4))
+- :sparkles: added md-to-json ([4bc0f9f](https://github.com/texttree/obs-format-convert-rcl/commit/4bc0f9f0c95f6f34a9832bf64ea3a9a939690f2e))
+- :sparkles: added md-to-zip ([1cab118](https://github.com/texttree/obs-format-convert-rcl/commit/1cab1181cc7c2595f5b0354082503677ab56f334))
+- :sparkles: added parameter that shows verse number ([c194bc3](https://github.com/texttree/obs-format-convert-rcl/commit/c194bc3313e794fad3a9c879556d1088cc773f4a))
+- :sparkles: json-to-pdf saves in batch ([1336a66](https://github.com/texttree/obs-format-convert-rcl/commit/1336a6652009c7c6a77103b6017743278f942962))
+- :sparkles: md-to-zip: batch save with folders ([800e9f1](https://github.com/texttree/obs-format-convert-rcl/commit/800e9f18c5581103c0a3e0b68b411fbef1a866af))
+
+### Bug Fixes
+
+- :bug: correction of remarks (path correction, description addition, config correction) ([1fe92a0](https://github.com/texttree/obs-format-convert-rcl/commit/1fe92a0a08f3537697489a2b6ae34539d2a6b133))
+- :bug: fix path ([8b7122f](https://github.com/texttree/obs-format-convert-rcl/commit/8b7122fb361be795c2c98fdd965d4a7dec2b2403))
+- :bug: fixed camelCase and imageUrl and added descriptions ([b35ff66](https://github.com/texttree/obs-format-convert-rcl/commit/b35ff66e9ffc485a189e6a83879b4da71caed60f))
+- :bug: fixed intro style redefinition ([19aeed6](https://github.com/texttree/obs-format-convert-rcl/commit/19aeed6a7cb13c5699d2a82234b2ae9bcee0ca69))
+- :bug: now, when combining verses, the style of verses does not disappear ([e859ab6](https://github.com/texttree/obs-format-convert-rcl/commit/e859ab690d12a7fd60623de5edbaf8ebf22493e1))
+- :bug: removing classes and adding style ([9cbb409](https://github.com/texttree/obs-format-convert-rcl/commit/9cbb409995d9d2bd4c1df36ed79d45f459af6a96))
+- :bug: style overriding disabled ([5a2cfc7](https://github.com/texttree/obs-format-convert-rcl/commit/5a2cfc7312a6ef8a5e61a7b4d6b3277688a5828a))
+
+### 0.1.1 (2023-05-22)